requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
python-dotenv==1.0.1